Reorganized to useroptions.py file
